+++ src/Exception/SmsException.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * This file is part of the "litgroup/sms" package.
+ *
+ * (c) Roman Shamritskiy <roman@litgroup.ru>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace LitGroup\Sms\Exception;
+
+/**
+ * Exception thrown if problem with short message service has occurred.
+ *
+ * @author Roman Shamritskiy <roman@litgroup.ru>
+ */
+class SmsException extends \Exception
+{
+    public function __construct($message, \Exception $previous = null)
+    {
+        parent::__construct($message, 0, $previous);
+    }
+}