+++ src/Sms.php
@@ -11,6 +11,7 @@
 namespace LitGroup\Sms;
 
 use LitGroup\Sms\Exception\GatewayException;
+use LitGroup\Sms\Exception\SmsException;
 use LitGroup\Sms\Gateway\GatewayInterface;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
@@ -53,11 +54,9 @@ class Sms implements SmsInterface
         try {
             $this->gateway->sendMessage($message);
         } catch (GatewayException $e) {
-            $this->logger->alert('Problem with SMS Gateway has occurred.', [
-                'exception' => $e
-            ]);
+            $this->logGatewayException($e);
 
-            throw $e;
+            throw new SmsException('Gateway problem has occurred.', $e);
         }
     }
 
@@ -69,5 +68,17 @@ class Sms implements SmsInterface
     private function setLogger(LoggerInterface $logger = null)
     {
         $this->logger = $logger !== null ? $logger : new NullLogger();
+    }
+
+    /**
+     * @param GatewayException $e
+     *
+     * @return void
+     */
+    private function logGatewayException(GatewayException $e)
+    {
+        $this->logger->alert('Problem with SMS Gateway has occurred.', [
+            'exception' => $e
+        ]);
     }
 }